Coverage make target (#41)
* Add a coverage make target.

Also detect when the coverage program is not installed and prompt the
user with an error.

* Also add program detection for autopep8 and pep8.

* Add detection for pyflakes as well.

* Switch to using flake8 for linting.

Flake8 wraps both pylint and pep8 in one call, simplifying the makefile
user experience down to just `make lint`.  It also automatically does
the scanning, removing the need for the find call and pipe to xargs.  It
also allows using #noqa to mark things that don't need to be checked, so
linting the files but ignoring unused imports in __init__ is supported.

- Adds lint target using flake8
- Adds check for presence of flake8 program
- Adds noqa lines in init files

* Ignore .coverage files.

* Conform to latest pep8.

Mostly just adds a space before the if __main__ blocks.

* Add notes about using the developers make targets.